Create 25.md
Posted article for 3/25/2025, but also added a "call to action" in the "one thing you can do today." Considering making it "my thing" to have the "OTAD" then directly below it, some quote or call to action. But for now, at least incorporating it into the OTAD helps the highlight the impact of the action. 
Note to self - consider specificity of the OTAD description. Is it general enough or too general?