Create 3.md
I think my tldr's are fine, and the actions that people who support or don't support the content in the articles are fine, but then for my OTAD call to action, sometimes it's hard for me to make it unbiased, like today's. Maybe more importantly than writing things unbiased, is to be clear/transparent about my/my tldr's bias for that day. One idea to do this is to either rank the tldrs on a scale of 5 (far left, left, central, right, far right), or maybe find a different way to scale it that doesn't include politics (such as a rating of 0-5 on how much my own bias may affect the article). I think that figuring out my main audience could help clear up this concern as well so I'll keep ruminating.